handle provider in state
