+++ extension/src/panel/app.tsx
@@ -1,59 +1,38 @@
 // This is the entrypoint to the panel app.
 // It has access to chrome.* APIs, but it can't interact with other extensions such as MetaMask.
-import { getChainId } from '@/chains'
 import { ZodiacToastContainer } from '@/components'
-import {
-  ProvideInjectedWallet,
-  ProvideProvider,
-  useProvider,
-} from '@/providers'
-import {
-  ProvideZodiacRoutes,
-  useConnectInjectedWalletIfNeeded,
-  useMarkRouteAsUsed,
-  useZodiacRoute,
-} from '@/zodiac-routes'
+import { ProvideInjectedWallet, ProvideProvider } from '@/providers'
+import { ProvideZodiacRoutes } from '@/zodiac-routes'
 import { invariant } from '@epic-web/invariant'
-import React, { useEffect } from 'react'
+import React from 'react'
 import { createRoot } from 'react-dom/client'
 import { createHashRouter, RouterProvider } from 'react-router-dom'
 import 'react-toastify/dist/ReactToastify.css'
-import { parsePrefixedAddress } from 'ser-kit'
-import { update } from '../inject/bridge'
 import { appRoutes } from './app-routes'
 import './global.css'
 import { initPort } from './port'
 import { ProvideState } from './state'
-import { useStorage } from './utils'
 
 initPort()
 
 const router = createHashRouter(appRoutes)
 
-const App = () => {
-  // update the last used timestamp for the current route
-  useMarkRouteAsUsed()
-
-  // make sure the injected provider stays updated on every relevant route change
-  const route = useZodiacRoute()
-
-  useConnectInjectedWalletIfNeeded(route)
-
-  const chainId = getChainId(route.avatar)
-  const provider = useProvider()
-  const [, avatarAddress] = parsePrefixedAddress(route.avatar)
-
-  useEffect(() => {
-    update(provider, chainId, avatarAddress)
-  }, [provider, chainId, avatarAddress])
-
-  useStorage('lastUsedRoute', route.id)
-
+const Root = () => {
   return (
-    <>
-      <RouterProvider router={router} />
-      <ZodiacToastContainer />
-    </>
+    <React.StrictMode>
+      <ProvideState>
+        <ProvideZodiacRoutes>
+          <ProvideInjectedWallet>
+            <ProvideProvider>
+              <div className="flex flex-1 flex-col">
+                <RouterProvider router={router} />
+                <ZodiacToastContainer />
+              </div>
+            </ProvideProvider>
+          </ProvideInjectedWallet>
+        </ProvideZodiacRoutes>
+      </ProvideState>
+    </React.StrictMode>
   )
 }
 
@@ -61,21 +40,7 @@ const rootEl = document.getElementById('root')
 
 invariant(rootEl != null, 'Could not find DOM node to attach app')
 
-createRoot(rootEl).render(
-  <React.StrictMode>
-    <ProvideState>
-      <ProvideZodiacRoutes>
-        <ProvideInjectedWallet>
-          <ProvideProvider>
-            <div className="flex flex-1 flex-col">
-              <App />
-            </div>
-          </ProvideProvider>
-        </ProvideInjectedWallet>
-      </ProvideZodiacRoutes>
-    </ProvideState>
-  </React.StrictMode>
-)
+createRoot(rootEl).render(<Root />)
 
 if (process.env.LIVE_RELOAD) {
   new EventSource(process.env.LIVE_RELOAD).addEventListener('change', (ev) => {
